Add unit tests project & implement create prep

+++ src/PrepApi/Contracts/CreateRecipeRequest.cs
@@ -18,7 +18,7 @@ public record CreateRecipeRequest
 public record RecipeIngredientInputDto
 {
     public required Guid IngredientId { get; init; }
-    public required double Quantity { get; init; }
+    public required decimal Quantity { get; init; }
     public required Unit Unit { get; init; }
 }
